refactor : 실패시 재시도 가능하게 수정 및 페이지 과부화 개선
