+++ src/main/java/nbc/chillguys/nzcrawler/review/crawler/DanawaReviewCrawler.java
@@ -6,8 +6,6 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Component;
 
-import com.microsoft.playwright.Browser;
-import com.microsoft.playwright.BrowserContext;
 import com.microsoft.playwright.ElementHandle;
 import com.microsoft.playwright.Locator;
 import com.microsoft.playwright.Page;
@@ -21,67 +19,61 @@ import nbc.chillguys.nzcrawler.review.dto.ReviewInfo;
 @Component
 public class DanawaReviewCrawler {
 
-	public List<ReviewInfo> crawl(Browser browser, Catalog catalog) {
+	public List<ReviewInfo> crawl(Page page, Catalog catalog) {
 		List<ReviewInfo> result = new ArrayList<>();
+		String url = "https://prod.danawa.com/info/?pcode=" + catalog.getProductCode();
 
-		try {
-			BrowserContext context = browser.newContext(new Browser.NewContextOptions()
-				.setUserAgent(getRandomUserAgent())
-				.setViewportSize(1280, 800)
-				.setLocale("ko-KR")
-				.setTimezoneId("Asia/Seoul")
-			);
+		for (int attempt = 1; attempt <= 3; attempt++) {
+			try {
+				page.navigate(url, new Page.NavigateOptions().setWaitUntil(WaitUntilState.DOMCONTENTLOADED));
+				page.addInitScript("Object.defineProperty(navigator, 'webdriver', {get: () => undefined})");
+				page.addInitScript("Object.defineProperty(navigator, 'languages', { get: () => ['ko-KR', 'ko'] });");
+				page.addInitScript("Object.defineProperty(navigator, 'platform', { get: () => 'Win32' });");
+				page.addInitScript("Object.defineProperty(navigator, 'plugins', { get: () => [1, 2, 3] });");
+				page.waitForTimeout(1000);
 
-			Page page = context.newPage();
+				Locator tab = page.locator("#danawa-prodBlog-productOpinion-button-tab-companyReview");
+				if (tab.count() == 0) {
+					log.warn("❌ [{}] 리뷰 탭 없음", catalog.getId());
+					return result;
+				}
+				tab.scrollIntoViewIfNeeded();
+				tab.click();
+				page.waitForTimeout(1000);
 
-			// navigator 속성 우회
-			page.addInitScript("Object.defineProperty(navigator, 'webdriver', {get: () => undefined})");
-			page.addInitScript("Object.defineProperty(navigator, 'languages', { get: () => ['ko-KR', 'ko'] });");
-			page.addInitScript("Object.defineProperty(navigator, 'platform', { get: () => 'Win32' });");
-			page.addInitScript("Object.defineProperty(navigator, 'plugins', { get: () => [1, 2, 3] });");
+				while (true) {
+					List<ElementHandle> items = page.querySelectorAll("li.danawa-prodBlog-companyReview-clazz-more");
+					if (items.isEmpty())
+						break;
 
-			String url = "https://prod.danawa.com/info/?pcode=" + catalog.getProductCode();
-			page.navigate(url, new Page.NavigateOptions().setWaitUntil(WaitUntilState.DOMCONTENTLOADED));
-			page.waitForTimeout(1000);
+					for (ElementHandle item : items) {
+						String content = Optional.ofNullable(item.querySelector(".atc"))
+							.map(ElementHandle::innerText)
+							.orElse(null);
+						int star = Optional.ofNullable(item.querySelector(".star_mask"))
+							.map(el -> parseStar(el.getAttribute("style")))
+							.orElse(0);
 
-			Locator tab = page.locator("#danawa-prodBlog-productOpinion-button-tab-companyReview");
-			if (tab.count() == 0) {
-				log.warn("❌ [{}] 리뷰 탭 없음", catalog.getId());
-				return result;
-			}
-			tab.scrollIntoViewIfNeeded();
-			tab.click();
-			page.waitForTimeout(1000);
+						if (content == null || content.isBlank())
+							continue;
+						result.add(new ReviewInfo(star, content.trim()));
+					}
 
-			while (true) {
-				List<ElementHandle> items = page.querySelectorAll("li.danawa-prodBlog-companyReview-clazz-more");
-				if (items.isEmpty())
-					break;
-
-				for (ElementHandle item : items) {
-					String content = Optional.ofNullable(item.querySelector(".atc"))
-						.map(ElementHandle::innerText)
-						.orElse(null);
-					int star = Optional.ofNullable(item.querySelector(".star_mask"))
-						.map(el -> parseStar(el.getAttribute("style")))
-						.orElse(0);
-
-					if (content == null || content.isBlank())
-						continue;
-
-					result.add(new ReviewInfo(star, content.trim()));
+					ElementHandle next = page.querySelector(".pagination .page_next");
+					if (next == null || !next.isVisible())
+						break;
+					next.click();
+					page.waitForTimeout(1500);
 				}
 
-				ElementHandle next = page.querySelector(".pagination .page_next");
-				if (next == null || !next.isVisible())
-					break;
-				next.click();
-				page.waitForTimeout(1500);
+				log.info("✅ [{}] 크롤링 성공 - {}개", catalog.getId(), result.size());
+				return result;
+			} catch (Exception e) {
+				log.warn("⏳ [{}] {}번째 시도 실패: {}", catalog.getId(), attempt, e.getMessage());
 			}
-			log.info("✅ [{}] 크롤링 성공 - {}개", catalog.getId(), result.size());
-		} catch (Exception e) {
-			log.warn("❌ [{}] 크롤링 실패: {}", catalog.getId(), e.getMessage());
 		}
+
+		log.error("❌ [{}] 최대 재시도 초과", catalog.getId());
 		return result;
 	}
 
@@ -91,13 +83,5 @@ public class DanawaReviewCrawler {
 		} catch (Exception e) {
 			return 0;
 		}
-	}
-
-	private String getRandomUserAgent() {
-		List<String> agents = List.of(
-			"Mozilla/5.0 (Windows NT 10.0; Win64; x64)...",
-			"Mozilla/5.0 (Macintosh; Intel Mac OS X)..."
-		);
-		return agents.get((int)(Math.random() * agents.size()));
 	}
 }
